requirements.txt
bottle==0.12.13
coincurve==8.0.0
dill==0.2.8.2
ecdsa==0.13
grequests==0.3.0
pycryptodomex==3.4.7
pyscrypt==1.6.2
PyYAML==3.12
pyzmq==17.0.0
requests==2.14.2
virtualenv==15.1.0
